requirements.txt
#json
jsonschema[format] >= 3.0.0b3
uritools
owlready2
pyxdg